Clean up _utils.TestCase.
- make it use the base ArmstrongTestCase
- move the setUp() and assertInContext() methods out of the base class and into the class they are used in
- remove an unnecessary WellViewTestCase class
