+++ armstrong/core/arm_wells/tests/_utils.py
@@ -1,12 +1,12 @@
 import random
 
-from django.test import TestCase as DjangoTestCase
-from django.test.client import RequestFactory
-
+from armstrong.dev.tests.utils.base import ArmstrongTestCase
 from .arm_wells_support.models import Story, StoryChild, Image
-from ..models import Node
-from ..models import Well
-from ..models import WellType
+from ..models import WellType, Well, Node
+
+
+class TestCase(ArmstrongTestCase):
+    pass
 
 
 def generate_random_image():
@@ -52,15 +52,3 @@ def generate_random_welltype():
     title = "Random Well %d" % r
     slug = "random-well-%d" % r
     return WellType.objects.create(title=title, slug=slug)
-
-
-class TestCase(DjangoTestCase):
-    def setUp(self):
-        self.factory = RequestFactory()
-
-    def assertInContext(self, var_name, other, template_or_context):
-        # TODO: support passing in a straight "context" (i.e., dict)
-        context = template_or_context.context_data
-        self.assertTrue(var_name in context,
-                msg="`%s` not in provided context" % var_name)
-        self.assertEqual(context[var_name], other)
