+++ armstrong/core/arm_wells/tests/arm_wells_support/tests.py
@@ -1,7 +1,6 @@
 import random
 
-from django.test import TestCase
-
+from .._utils import TestCase
 from .models import Story
